+++ CHANGES.rst
@@ -1,5 +1,12 @@
-Release 0.3.0
-=============
+Release 0.3
+============
+
+Major changes
+-------------
+
+* New encoder: :class:`DatetimeEncoder` can transform a datetime column into several numerical
+    columns (year, month, day, hour, minute, second, ...). It is now the default transformer used
+    in the SuperVectorizer for datetime columns.
 
 Notes
 -----